requirements.txt
requests>=2.31.0
geopandas>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
shapely>=2.0.0
folium>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
networkx>=3.1.0
overpy>=0.7.0
rasterio>=1.3.0
elevation>=1.1.0
osmnx>=1.6.0
geopy>=2.3.0
tqdm>=4.65.0
pyyaml>=6.0.0
click>=8.1.0
flask>=2.3.0
jupyter>=1.0.0
plotly>=5.15.0
dash>=2.11.0
python-dotenv>=1.0.0